requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asgiref==3.7.2
attr==0.3.1
attrs==23.2.0
azure-core==1.29.7
azure-storage-blob==12.19.0
black==24.1.1
bleach==5.0.1
blinker==1.7.0
boto==2.49.0
boto3==1.34.30
botocore==1.34.30
boxing==0.1.4
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@b7ff9466d174fbb7061ff6d3773cd9c372a8e56f
Django==3.2.23
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
easyocr==1.7.1
et-xmlfile==1.1.0
expiringdict==1.2.2
filelock==3.13.1
Flask==3.0.1
Flask-Reuploaded==1.4.0
Flask-WTF==1.2.1
fonttools==4.47.2
fsspec==2023.12.2
fvcore==0.1.5.post20221221
google-api-core==2.16.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
htmlmin==0.1.12
huggingface-hub==0.20.3
humansignal-drf-yasg==1.21.9
hydra-core==1.3.2
idna==3.6
ijson==3.2.3
imageio==2.33.1
inflection==0.5.1
iopath==0.1.9
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==3.2.0
kiwisolver==1.4.5
label-studio==1.10.1
label-studio-converter==0.0.57
label-studio-tools==0.0.3
launchdarkly-server-sdk==7.5.0
lazy_loader==0.3
lockfile==0.12.2
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.2
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.6.7
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
ordered-set==4.0.2
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
pdf2image==1.17.0
pillow==10.2.0
pip-system-certs==4.0
platformdirs==4.1.0
portalocker==2.8.2
protobuf==4.23.4
psutil==5.9.8
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.14
pyparsing==3.1.1
pyRFC3339==1.1
pyrsistent==0.20.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7.1
pywin32==306
PyYAML==6.0.1
redis==3.5.3
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.0
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
seaborn==0.13.2
semver==2.13.0
sentry-sdk==1.39.2
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.1.30
tokenizers==0.15.1
tomli==2.0.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2023.4
ua-parser==0.18.0
ujson==5.9.0
ultralytics==8.1.15
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.16.0
WTForms==3.1.2
xmljson==0.2.0
yacs==0.1.8
